requirements.txt
# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0
PyMuPDF>=1.23.0  # PyMuPDF for advanced PDF processing

# XML Generation
lxml>=4.9.0
xmlschema>=2.5.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0

# Machine Learning (optional)
scikit-learn>=1.3.0
transformers>=4.21.0

# Web Interface
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# CLI
click>=8.1.0
rich>=13.0.0
typer>=0.9.0

# Configuration
pydantic>=2.0.0
python-decouple>=3.8

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0